four btns

+++ firebase.js
@@ -1,6 +1,6 @@
 import { initializeApp } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-app.js";
 import { getAuth ,createUserWithEmailAndPassword ,signInWithEmailAndPassword,onAuthStateChanged,sendEmailVerification ,updateProfile, signOut ,GoogleAuthProvider,signInWithPopup } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-auth.js";
-import {getFirestore,collection, addDoc , doc, setDoc , getDocs,deleteDoc,updateDoc } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-firestore.js";
+import {getFirestore,collection, addDoc , doc, setDoc , getDocs,deleteDoc,updateDoc, serverTimestamp,getDoc } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-firestore.js";
 
 
 
@@ -20,5 +20,8 @@ const app = initializeApp(firebaseConfig);
 const provider = new GoogleAuthProvider();
 const db = getFirestore(app);
 
-export {getAuth,createUserWithEmailAndPassword, signInWithEmailAndPassword ,onAuthStateChanged,sendEmailVerification,updateProfile, signOut,GoogleAuthProvider,provider,signInWithPopup,db,collection, addDoc , doc, setDoc , getDocs,deleteDoc,updateDoc}
+const auth = getAuth();
+const user = auth.currentUser;
+
+export {getAuth,createUserWithEmailAndPassword, signInWithEmailAndPassword ,onAuthStateChanged,sendEmailVerification,updateProfile, signOut,GoogleAuthProvider,provider,signInWithPopup,db,collection, addDoc , doc, setDoc , getDocs,deleteDoc,updateDoc,user, serverTimestamp,getDoc,auth}
